+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 - Switch: Fixed disabled label color
 - Image: Added support for passing attributes.style
 - Modal, Overlay: Added onClose 'reason'
+- useResponsiveClientValue: boolean values support